# go/model_list_request.go
/*
 * Rosberry Mobile APP API (Test work)
 *
 * Test work for rosberry
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ListRequest struct {

	UserId int32 `json:"userId,omitempty"`

	Limit int32 `json:"limit,omitempty"`

	Offset int32 `json:"offset,omitempty"`
}