# go/model_profile.go
/*
 * Rosberry Mobile APP API (Test work)
 *
 * Test work for rosberry
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Profile struct {

	UserId int32 `json:"userId,omitempty"`

	Name string `json:"name,omitempty"`

	// image data in base64
	Avatar string `json:"avatar,omitempty"`

	// birthday in ISO 8601
	Birthday string `json:"birthday,omitempty"`

	Interests []string `json:"interests,omitempty"`
}